Move ObsoleteTypeHintInspectionTests for QuickFixes into relevant files under RD.QuickFixes
